requirements.txt
streamlit
crewai-tools
crewai
python-dotenv
